backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
google-genai>=1.14.0
openpyxl>=3.1.5
Pillow>=11.0.0
pydantic>=2.11.0
pydantic-settings>=2.8.0
python-dotenv>=1.0.1
httpx>=0.28.0
aiofiles>=24.1.0
sqlalchemy>=2.0.0
websockets>=13.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
